Add option to not require root permissions
If the -n|--noroot option is passed, the script will not assume that root permissions are
needed to write the output file. The script will fail if the -n|--noroot option is passed
along with the default output location.

This will allow custom output locations to not ask for sudo permission if it is not needed.