Add hosts-file.net ad list to default sources
While there may be more domains on the hosts-file ad list, it is a large encompassing list that
can catch addresses that would otherwise fall through. Preserving the unbreak whitelist should
reduce the issues encountered. If issues occur, please report to this project instead of uBlock0
unless the issue is caused from a lack of whitelist entry