2.2.0 does not include unbreak list by default
The unbreak list is only good when the user has a second line of defense in place such as
uBlock0. When there is no other blocker, the unbreak whitelist may leak potential identifying
information to these otherwise blacklisted sites.

Blacklist all sites by default, and allow the user to include the unbreak list when needed.
The unbreak list will still be updated as regularly as possible.